+++ apps/passport-server/src/services/generic-issuance/GenericIssuanceService.ts
@@ -8,11 +8,13 @@ import {
   GenericIssuanceSemaphoreGroupResponseValue,
   GenericIssuanceSemaphoreGroupRootResponseValue,
   GenericIssuanceSendEmailResponseValue,
+  GenericIssuanceSendPipelineEmailResponseValue,
   GenericIssuanceValidSemaphoreGroupResponseValue,
   GenericPretixEvent,
   GenericPretixProduct,
   ListFeedsResponseValue,
   PipelineDefinition,
+  PipelineEmailType,
   PipelineGetManualCheckInsResponseValue,
   PipelineInfoResponseValue,
   PipelineSetManualCheckInStateResponseValue,
@@ -38,6 +40,10 @@ import {
   PipelineConsumerDB
 } from "../../database/queries/pipelineConsumerDB";
 import {
+  IPipelineEmailDB,
+  PipelineEmailDB
+} from "../../database/queries/pipelineEmailDB";
+import {
   IPipelineManualTicketDB,
   PipelineManualTicketDB
 } from "../../database/queries/pipelineManualTicketDB";
@@ -54,6 +60,7 @@ import {
 import { ApplicationContext } from "../../types";
 import { logger } from "../../util/logger";
 import { DiscordService } from "../discordService";
+import { EmailService } from "../emailService";
 import { PagerDutyService } from "../pagerDutyService";
 import { PersistentCacheService } from "../persistentCacheService";
 import { InMemoryPipelineAtomDB } from "./InMemoryPipelineAtomDB";
@@ -78,6 +85,7 @@ export class GenericIssuanceService {
   private checkinDB: IPipelineCheckinDB;
   private contactDB: IContactSharingDB;
   private badgeDB: IBadgeGiftingDB;
+  private emailDB: IPipelineEmailDB;
   private consumerDB: IPipelineConsumerDB;
   private manualTicketDB: IPipelineManualTicketDB;
   private semaphoreHistoryDB: IPipelineSemaphoreHistoryDB;
@@ -86,6 +94,7 @@ export class GenericIssuanceService {
   private pipelineSubservice: PipelineSubservice;
   private userSubservice: UserSubservice;
   private credentialSubservice: CredentialSubservice;
+  private emailService: EmailService;
 
   public constructor(
     context: ApplicationContext,
@@ -98,6 +107,7 @@ export class GenericIssuanceService {
     rollbarService: RollbarService | null,
     pagerdutyService: PagerDutyService | null,
     discordService: DiscordService | null,
+    emailService: EmailService,
     cacheService: PersistentCacheService
   ) {
     this.context = context;
@@ -109,7 +119,9 @@ export class GenericIssuanceService {
     this.genericPretixAPI = pretixAPI;
     this.contactDB = new ContactSharingDB(this.context.dbPool);
     this.badgeDB = new BadgeGiftingDB(this.context.dbPool);
+    this.emailDB = new PipelineEmailDB(this.context.dbPool);
     this.pipelineAtomDB = new InMemoryPipelineAtomDB();
+    this.emailService = emailService;
     this.userSubservice = new UserSubservice(
       context,
       stytchClient,
@@ -136,11 +148,13 @@ export class GenericIssuanceService {
         pipelineAtomDB: this.pipelineAtomDB,
         checkinDB: this.checkinDB,
         contactDB: this.contactDB,
+        emailDB: this.emailDB,
         badgeDB: this.badgeDB,
         consumerDB: this.consumerDB,
         manualTicketDB: this.manualTicketDB,
         semaphoreHistoryDB: this.semaphoreHistoryDB,
         credentialSubservice: this.credentialSubservice,
+        emailService,
         context
       } satisfies InstantiatePipelineArgs
     );
@@ -319,6 +333,13 @@ export class GenericIssuanceService {
     pipelineId: string
   ): Promise<GenericIssuancePipelineSemaphoreGroupsResponseValue> {
     return this.pipelineSubservice.handleGetPipelineSemaphoreGroups(pipelineId);
+  }
+
+  public async handleSendPipelineEmail(
+    pipelineId: string,
+    email: PipelineEmailType
+  ): Promise<GenericIssuanceSendPipelineEmailResponseValue> {
+    return this.pipelineSubservice.handleSendPipelineEmail(pipelineId, email);
   }
 
   /**